Refactor Ffprobe controllers
Testing Ffprobe's output is not a good idea, as it can be different for each Bluray disc, depending streams' codecs. Hence, the Ffprobe controllers and Bluray analyzer have been refactored. Now, the analyzer is also tested in integration tests, in order to check its behavior with either the real or stub Ffprobe controller.